+++ ProvaEDII/Prova1/src/AlgoritmosOrdenacao/MetricasAlgoritmos.java
@@ -1,7 +1,7 @@
 package AlgoritmosOrdenacao;
 
 public class MetricasAlgoritmos {
-	public static void metricas(long tempo, int comparacoes, int trocas) {
+	public static void metricas(long tempo, long comparacoes, int trocas) {
 		long tempoEmSeg = tempo/1000;
 		
 		System.out.println("MÉTRICAS:\nTEMPO DE EXECUÇÃO: "+tempoEmSeg+"s\nTOTAL DE COMPARAÇÕES: "+comparacoes+"\nTOTAL DE TROCAS: "+trocas);
